Saturday final push

+++ avionics_firmware/data_buffer.h
@@ -18,10 +18,10 @@
 
 // Define Constants and Thresholds
 #define BUFFER_SIZE       50
-#define LAUNCH_THRESHOLD  50      // micro bar for detecting a decrease
+#define LAUNCH_THRESHOLD  50     // micro bar for detecting a decrease
 #define APOGEE_THRESHOLD  50      // micro bar for detecting apogee
 #define DESCENT_THRESHOLD 50      // micro bar for detecting an increase
-#define GROUND_THRESHOLD  100      // micro bar for detecting ground
+#define GROUND_THRESHOLD  100     // micro bar for detecting ground
 #define WINDOW_SIZE       20      // Number of readings to compute
 
 static float sea_level_pressure = 1013.25;    // Sea level presser in micro bar
@@ -63,7 +63,7 @@ void update_buffer(FrameArray* frame, dataBuffer* buffer);
   @param dt - time interval between readings in microseconds
   @return float vertical velocity value in m/s
 */
-float get_vertical_velocity(int data[], int size, int dt);
+int get_vertical_velocity(int data[], int size, int dt);
 
 /**
   @brief Check if the rocket is stationary
